Summer 2021 Update
Compatible with AUX_MotherBoard V4/V5 Code

This commit features a new, separate code for the lights board, Aux_LightsBoard. It is designed to work with the Summer 2021 AUX_MotherBoard code.

The old code for the lights board is SolarGators_Auxiliary_SteeringWheelPROTOTYPE_V0. The code was last worked on by Aidan and makes use of the Solar Gators CAN drivers. The new code does not.